refactor: move provider onboarding into extensions
