+++ src/commands/onboard-auth.ts
@@ -1,8 +1,5 @@
-export {
-  SYNTHETIC_DEFAULT_MODEL_ID,
-  SYNTHETIC_DEFAULT_MODEL_REF,
-} from "../agents/synthetic-models.js";
-export { VENICE_DEFAULT_MODEL_ID, VENICE_DEFAULT_MODEL_REF } from "../agents/venice-models.js";
+export { SYNTHETIC_DEFAULT_MODEL_ID } from "../agents/synthetic-models.js";
+export { VENICE_DEFAULT_MODEL_ID } from "../agents/venice-models.js";
 export {
   applyAuthProfileConfig,
   applyCloudflareAiGatewayConfig,
@@ -91,15 +88,17 @@ export {
   setXaiApiKey,
   setModelStudioApiKey,
   writeOAuthCredentials,
-  HUGGINGFACE_DEFAULT_MODEL_REF,
   VERCEL_AI_GATEWAY_DEFAULT_MODEL_REF,
   XIAOMI_DEFAULT_MODEL_REF,
   ZAI_DEFAULT_MODEL_REF,
-  TOGETHER_DEFAULT_MODEL_REF,
-  MISTRAL_DEFAULT_MODEL_REF,
-  XAI_DEFAULT_MODEL_REF,
   MODELSTUDIO_DEFAULT_MODEL_REF,
 } from "./onboard-auth.credentials.js";
+export { HUGGINGFACE_DEFAULT_MODEL_REF } from "../../extensions/huggingface/onboard.js";
+export { MISTRAL_DEFAULT_MODEL_REF } from "../../extensions/mistral/onboard.js";
+export { SYNTHETIC_DEFAULT_MODEL_REF } from "../../extensions/synthetic/onboard.js";
+export { TOGETHER_DEFAULT_MODEL_REF } from "../../extensions/together/onboard.js";
+export { VENICE_DEFAULT_MODEL_REF } from "../../extensions/venice/onboard.js";
+export { XAI_DEFAULT_MODEL_REF } from "../../extensions/xai/onboard.js";
 export {
   buildKilocodeModelDefinition,
   buildMinimaxApiModelDefinition,
